IOS-5119 Replace Int with UInt64 for eth fee

+++ BlockchainSdk/WalletManagers/Ethereum/EthereumWalletManager.swift
@@ -133,7 +133,8 @@ private extension EthereumWalletManager {
 
                     // TODO: Fix integer overflow. Think about BigInt
                     // https://tangem.atlassian.net/browse/IOS-4268
-                    let fee = Decimal(Int(feeValue)) / decimalValue
+                    // https://tangem.atlassian.net/browse/IOS-5119
+                    let fee = Decimal(UInt64(feeValue)) / decimalValue
 
                     let amount = Amount(with: blockchain, value: fee)
                     let parameters = EthereumFeeParameters(gasLimit: gasLimit, gasPrice: gasPrice)
